simplify dockerfile around the python:3.8 image
we can run this service using the official Python docker images, which
gets us tools like `pip` out of the box. By removing build
instructions that weren't necessary we can dramatically reduce the
complexity of this file, and by restructuring it we can get faster
builds because we are able to take more advantage of the cache.

This commits only touches Dockerfile.x86.