add docker post_push hook
set up a post_push hook as described in [1] so that when running
automatic builds on docker hub (builds triggered from git commits), we
generate an image tagged with the commit.

this makes it very easy to pin a deployment to a specific commit (to
avoid breaking something if the `latest` tag changes), and also makes
it easy to force a deployment refresh when a new image is available
(because changing the image tag in the deployment configuration will
cause it to refresh the running pod(s).

[1] https://docs.docker.com/docker-hub/builds/advanced/